refactor: FileIOException 패키지 위치 변경
member, voucher에서 모두 사용가능하도록 변경
