+++ src/main/java/com/programmers/vouchermanagement/exception/FileIOException.java
@@ -1,4 +1,4 @@
-package com.programmers.vouchermanagement.voucher.exception;
+package com.programmers.vouchermanagement.exception;
 
 public class FileIOException extends RuntimeException {
